refactor: Move interfaces
